requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
retell-sdk==4.36.0
twilio==8.10.0
plivo==4.57.0
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.2
phonenumbers==8.13.26
pytz==2023.3